Added the doc and cleared unnecessary files. Reorganized doc
