+++ crypto_history/__init__.py
@@ -1,4 +1,5 @@
 from .utilities.general_utilities import init_logger, AbstractFactory  # NOQA
-from .core import stock_market_factory, data_container  # NOQA
+from .data_container import data_container_intra  # NOQA
+from .stock_market import stock_market_factory
 
 class_builders = AbstractFactory.get_builders()
