+++ crypto_history/version.py
@@ -1,1 +1,1 @@
-__version__ = "1.1.a3"
+__version__ = "1.2.a0"
