+++ docs/index.rst
@@ -10,13 +10,13 @@ Welcome to crypto_history's documentation!
 .. toctree::
    :caption: Source Code
 
-   source_code/market_data
-   source_code/data_container
+   source_code/market/stock_market_factory
+   source_code/data_container/data_container_pre
+   source_code/data_container/data_container_intra
    source_code/utilities
 
 Indices and tables
 ==================
 
 * :ref:`genindex`
-* :ref:`modindex`
 * :ref:`search`
