--- docs/source_code/data_container.rst
@@ -1,37 +1,0 @@
-Data Container
-===============
-
-The following provide information on storing the market data in a data container.
-
-
-.. autoclass:: crypto_history.core.data_container.DataFactory()
-    :members:
-    :undoc-members:
-
-.. autoclass:: crypto_history.core.data_container.ConcreteXArrayFactory()
-    :members:
-    :undoc-members:
-
-.. autoclass:: crypto_history.core.data_container.AbstractCoinHistoryObtainer()
-    :members:
-    :undoc-members:
-
-.. autoclass:: crypto_history.core.data_container.XArrayCoinHistoryObtainer()
-    :members:
-    :undoc-members:
-
-.. autoclass:: crypto_history.core.data_container.AbstractDataContainerOperations()
-    :members:
-    :undoc-members:
-
-.. autoclass:: crypto_history.core.data_container.XArrayDataContainerOperations()
-    :members:
-    :undoc-members:
-
-.. autoclass:: crypto_history.core.data_container.AbstractDimensionsManager()
-    :members:
-    :undoc-members:
-
-.. autoclass:: crypto_history.core.data_container.XArrayDimensionsManager()
-    :members:
-    :undoc-members:
