+++ docs/source_code/data_container/data_container_intra.rst
@@ -1,0 +1,13 @@
+Data Container
+===============
+
+The following provide information on how the high level operations on the data container are performed
+
+
+.. autoclass:: crypto_history.data_container.data_container_intra.TimeStampIndexedDataContainer()
+    :members:
+    :undoc-members:
+
+.. autoclass:: crypto_history.data_container.data_container_intra.TimeAggregatedDataContainer()
+    :members:
+    :undoc-members:
