+++ docs/source_code/data_container/data_container_pre.rst
@@ -1,0 +1,22 @@
+Primitive Data Container
+========================
+
+The following provide information on how the primitive data array is built
+
+
+.. autoclass:: crypto_history.data_container.data_container_pre.PrimitiveCoinHistoryObtainer()
+    :members:
+    :undoc-members:
+
+.. autoclass:: crypto_history.data_container.data_container_pre.PrimitiveDimensionsManager()
+    :members:
+    :undoc-members:
+
+.. autoclass:: crypto_history.data_container.data_container_pre.PrimitiveDataArrayOperations()
+    :members:
+    :undoc-members:
+
+
+.. autoclass:: crypto_history.data_container.utilities.DataFrameOperations()
+    :members:
+    :undoc-members:
