+++ docs/source_code/market/stock_market_factory.rst
@@ -1,0 +1,57 @@
+Market Data
+=============
+
+The following provide information on obtaining market data from crypto-currency exchanges and
+market APIs.
+
+.. autoclass:: crypto_history.stock_market.stock_market_factory.StockMarketFactory()
+    :members:
+    :undoc-members:
+
+.. autoclass:: crypto_history.stock_market.stock_market_factory.ConcreteBinanceFactory()
+    :members:
+    :undoc-members:
+
+
+.. autoclass:: crypto_history.stock_market.stock_market_factory.AbstractMarketHomogenizer()
+    :members:
+    :undoc-members:
+
+.. autoclass:: crypto_history.stock_market.stock_market_factory.BinanceHomogenizer()
+    :members:
+    :undoc-members:
+
+
+.. autoclass:: crypto_history.stock_market.stock_market_factory.AbstractMarketOperations()
+    :members:
+    :undoc-members:
+
+.. autoclass:: crypto_history.stock_market.stock_market_factory.BinanceMarketOperations()
+    :members:
+    :undoc-members:
+
+
+.. autoclass:: crypto_history.stock_market.request.AbstractMarketRequester()
+    :members:
+    :undoc-members:
+
+.. autoclass:: crypto_history.stock_market.request.BinanceRequester()
+    :members:
+    :undoc-members:
+
+
+.. autoclass:: crypto_history.stock_market.stock_market_factory.AbstractOHLCVFieldTypes()
+    :members:
+    :undoc-members:
+
+.. autoclass:: crypto_history.stock_market.stock_market_factory.BinanceRequester()
+    :members:
+    :undoc-members:
+
+.. autoclass:: crypto_history.stock_market.stock_market_factory.AbstractTimeIntervalChunks()
+    :members:
+    :undoc-members:
+
+.. autoclass:: crypto_history.stock_market.stock_market_factory.BinanceTimeIntervalChunks()
+    :members:
+    :undoc-members:
