--- docs/source_code/market_data.rst
@@ -1,41 +1,0 @@
-Market Data
-=============
-
-The following provide information on obtaining market data from crypto-currency exchanges and
-market APIs.
-
-.. autoclass:: crypto_history.core.get_market_data.AbstractMarketHomogenizer()
-    :members:
-    :undoc-members:
-
-.. autoclass:: crypto_history.core.get_market_data.BinanceHomogenizer()
-    :members:
-    :undoc-members:
-    :exclude-members: HistoryFields
-
-
-.. autoclass:: crypto_history.core.get_market_data.AbstractMarketOperations()
-    :members:
-    :undoc-members:
-
-.. autoclass:: crypto_history.core.get_market_data.BinanceMarketOperations()
-    :members:
-    :undoc-members:
-
-
-.. autoclass:: crypto_history.core.get_market_data.StockMarketFactory()
-    :members:
-    :undoc-members:
-
-.. autoclass:: crypto_history.core.get_market_data.ConcreteBinanceFactory()
-    :members:
-    :undoc-members:
-
-
-.. autoclass:: crypto_history.core.request.AbstractMarketRequester()
-    :members:
-    :undoc-members:
-
-.. autoclass:: crypto_history.core.request.BinanceRequester()
-    :members:
-    :undoc-members:
